move FermoinMapBoson class to qchem.py
